+++ Umbraco.Plugins.Yaml2Schema/src/Composers/YamlStartupComposer.cs
@@ -27,6 +27,8 @@ namespace Umbraco.Plugins.Yaml2Schema.Composers
             builder.Services.AddScoped<DictionaryCreator>();
             builder.Services.AddScoped<MemberCreator>();
             builder.Services.AddScoped<UserCreator>();
+            builder.Services.AddScoped<PackageValidator>();
+            builder.Services.AddScoped<PropertyEditorCreator>();
 
             // IHttpClientFactory is needed by MediaCreator for URL-based file downloads
             builder.Services.AddHttpClient();